+++ pi-monitor/pi-monitor-client/sdks/core/src/commonMain/kotlin/pimonitor/PiMonitorService.kt
@@ -3,11 +3,9 @@
 package pimonitor
 
 import bitframe.BitframeService
-import bitframe.authentication.ClientConfiguration
 import bitframe.authentication.signin.SignInService
 import pimonitor.authentication.signup.SignUpService
 import pimonitor.evaulation.business.BusinessService
-import platform.ExecutionEnvironment
 import kotlin.js.JsExport
 
 abstract class PiMonitorService : BitframeService {